Revert "Merge pull request #5 from codequest07/amd-updates"
This reverts commit 6adff535085891a08513c5c727eebf8d494b485e, reversing
changes made to 7941995c74bba1ddbb51835cd2e6823ea4563be1.
